Progress with new RTCC coordinate systems
