+++ Orbitersdk/samples/ProjectApollo/src_rtccmfd/OrbMech.h
@@ -198,75 +198,6 @@ struct LGCIgnitionConstants
 	double K_V = 410.0;
 };
 
-struct AEGHeader
-{
-	//0 = Earth, 1 = Lunar
-	int AEGInd = 0;
-	int ErrorInd = 0;
-	int NumBlocks = 1;
-	//int Spare;
-};
-
-struct AEGDataBlock
-{
-	//Item 1, 11 is Keplerian to Keplerian (only used in AEG service routine)
-	int InputOutputInd = 11;
-	//Item 2, initialize/update indicator (0 = Osculating elements provided, 1 = osc and mean elements provided, use low e form., 2 = same as 1, but high e)
-	int ENTRY = 0;
-	//Item 3, update option indicator (0 = to time, 1 = to mean anomaly, 2 = to argument of latitude, 3 = to maneuver counter line...)
-	int TIMA = 0;
-	//Item 4, 0 for J2 and J4, 1 for J2, J3, J4 and -1 for J2 and J3 (only 1 should be used in the real time system)
-	int HarmonicsInd = 1;
-	//Item 5, drag indicator. If nonzero it is the K-Factor
-	double ICSUBD = 0.0;
-	//Item 6, vehicle area
-	double VehArea = 0.0;
-	//Item 7, vehicle weight for Earth AEG, GMTBASE for Lunar AEG
-	double Item7;
-	//Item 8, Input: Mean anomaly (option 1), argument of latitude (option 2-3), Output: DH (options 5-6) otherwise same as input
-	double Item8;
-	//Item 9, crossing time of the reference counter line (input), time lag (options 5-6), otherwise same as input (output)
-	double Item9;
-	//Item 10, number of times the AEG will cross the reference line (input), phase angle for options 4-6 (output)
-	double Item10;
-	CELEMENTS coe_osc;
-	CELEMENTS coe_mean;
-	//Item 23, time associated with elements 11-16 (in GMT)
-	double TS;
-	double l_dot;
-	double g_dot;
-	double h_dot;
-	//Item 27, time to which to update elements (in GMT)
-	double TE;
-	double f;
-	double U;
-	double R;
-};
-
-struct AEGBlock
-{
-	AEGHeader Header;
-	AEGDataBlock Data;
-};
-
-class PMMAEG
-{
-public:
-	PMMAEG();
-	void CALL(AEGHeader &header, AEGDataBlock &in, AEGDataBlock &out);
-protected:
-	AEGDataBlock CurrentBlock;
-};
-
-class PMMLAEG
-{
-public:
-	PMMLAEG();
-	void CALL(AEGHeader &header, AEGDataBlock &in, AEGDataBlock &out);
-protected:
-	AEGDataBlock CurrentBlock;
-};
-
 class CoastIntegrator
 {
 public:
